Improve post GUID fixup logic
Previously, the update logic would run every time for a post with
an `&` in its GUID, even if the DB had been fixed.  Also, when the
db needed to be fixed, the WP post cache would still contain the
incorrect value.  (Also, the post_type field was unnecessarily being
updated, as it was only needed for the PostModel::on_save_post()
call.)

As soon as practical, this entire code chunk should move to
Imposer's PostModel, both because it depends on PostModel internals
and because the GUID issue affects anything that wants to impose a
post's GUID, not just Postmark.